+++ extensions/confluence-search/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Confluence Search Changelog
 
+ ## [Update] - 2023-02-09
+ - feat: add confluence filter by type, support page, blog, attachment.
+ - feat: default only search page, [#4340](https://github.com/raycast/extensions/issues/4340)
+ - fix: [#4523](https://github.com/raycast/extensions/issues/4523)
+
  ## [Update] - 2023-02-07
 
  - fix: fix user icon not working